Final version before major redesign #2
This is the final version prior to a major redesign.

If has one known issue for calculating TC for Juliet False for CWE 562 only (incorrectly is zero vs 1. This is because there are no 'path' values for this project and does not count correctly in the current configuration. Other Juliet false TC have not been verified.

All other counts and hits are believed to be correct.